handlers/start_handler.py
import logging
from telegram import Update
from telegram.ext import CommandHandler, ContextTypes

logger = logging.getLogger(__name__)


async def hello(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await context.bot.send_message(chat_id=update.effective_chat.id, text="Hello, This is the signal swap bot, please explore! Please send command /menu to do more things.")

handler = CommandHandler('start', hello)