pyproject.toml
[tool.poetry]
name = "tg-bot-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.6"
eth-account = "^0.9.0"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
